Implement #check if file exists before delete

+++ src/CloudMunch/CloudmunchService.php
@@ -298,7 +298,9 @@ class CloudmunchService {
 	 */
 	public function deleteKeys() {
 		foreach ( $this->keyArray as $file ) {
-			system ( "rm " . $file );
+			if(file_exists($file)){
+				system ( "rm " . $file );
+			}
 		}
 	}
 }
